initial commit
Someday I'd like to move my mainstream blog to Pelican and have all
the posts in Git; right now it's still WordPress (because that was the
obvious thing to do when I set it up in 2011). But before I've moved
the whole blog, I still want new posts in version control, so let's
init a repository for that.