tap "College" ... before an executive decision
You know what? It's 31 December. I was imagining finishing the
prologue, but in addition to it being expensive to confront the past,
it's also plausibly not the highest value task. It's not "now or
never"; I could pull it out into a prequel post, and having this post
be narrowly focused on my return and triumph at SFSU is more cohesive.